# reactome_analysis_api/reactome_analysis_api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a Gene Set Analysis (GSA) Service, which provides GSA methods\
    \  whose results can be visualized using the Reactome pathway browser."
  version: "0.1.0"
  title: "Reactome Gene Set Analysis Service"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/0.1"
...
schemes:
- "https"
- "http"
paths:
...
  /analysis:
    post:
      tags:
      - "Analysis"
      summary: "Performs the specified gene set analysis"
      operationId: "start_analysis"
      consumes:
      - "application/json"
      - "application/gzip"
      produces:
      - "text/plain; charset=utf-8"
      parameters:
      - in: "body"
        name: "body"
        description: "Specification of analysis to perform"
        required: true
        schema:
          $ref: "#/definitions/AnalysisInput"
      responses:
        200:
          description: "Successful start of the analysis. Returns the identifier of\
            \ the running analysis."
          schema:
            type: "string"
            example: "Analysis00371643"
        400:
          description: "JSON request does not follow the specification."
        404:
          description: "Unknown analysis method selected."
        406:
          description: "Invalid analysis specifcation submitted."
        503:
          description: "The internal queuing or storage system are currently unavailable.\
            \ This is generally due to an ongoing update. Please retry in 4 seconds."
      x-swagger-router-controller: "reactome_analysis_api.controllers.analysis_controller"
...
definitions:
...
  AnalysisInput:
    type: "object"
    required:
    - "datasets"
    - "methodName"
    properties:
      methodName:
        type: "string"
        example: "Camera"
      datasets:
        type: "array"
        items:
          $ref: "#/definitions/Dataset"
      parameters:
        type: "array"
        items:
          $ref: "#/definitions/Parameter"
      analysisId:
        type: "string"
        description: "This field is only used internally and will be ignored if set\
          \ in the request."
    example:
      methodName: "Camera"
      analysisId: "analysisId"
      datasets:
      - data: "\tSample 1\tSample2\tSample 3\nCD19\t10\t20\t5\nCD20\t4\t7\t30\nMITF\t\
          3\t9\t20\nPTEN\t10\t4\t20\n"
        design:
          comparison:
            group2: "Treatment"
            group1: "Control"
          analysisGroup:
          - "Treatment"
          - "Control"
          - "Treatment"
          samples:
          - "Sample 1"
          - "Sample 2"
          - "Sample 3"
        name: "First experiment"
        type: "rnaseq_counts"
        parameters:
        - name: "permutations"
          value: "10"
        - name: "permutations"
          value: "10"
      - data: "\tSample 1\tSample2\tSample 3\nCD19\t10\t20\t5\nCD20\t4\t7\t30\nMITF\t\
          3\t9\t20\nPTEN\t10\t4\t20\n"
        design:
          comparison:
            group2: "Treatment"
            group1: "Control"
          analysisGroup:
          - "Treatment"
          - "Control"
          - "Treatment"
          samples:
          - "Sample 1"
          - "Sample 2"
          - "Sample 3"
        name: "First experiment"
        type: "rnaseq_counts"
        parameters:
        - name: "permutations"
          value: "10"
        - name: "permutations"
          value: "10"
      parameters:
      - name: "permutations"
        value: "10"
      - name: "permutations"
        value: "10"
  Dataset:
    type: "object"
    required:
    - "data"
    - "name"
    - "type"
    properties:
      name:
        type: "string"
        example: "First experiment"
        description: "Every dataset must have a unique name."
      type:
        type: "string"
        example: "rnaseq_counts"
        description: "Specifies the type of dataset. Currently supported types are\
          \ RNA-seq (raw read counts), intensity-based proteomics quantification (proteomics-int),\
          \ raw proteomics spectral counts (proteomics-sc), and microarray data."
        enum:
        - "rnaseq_counts"
        - "rnaseq_norm"
        - "proteomics_int"
        - "proteomics_sc"
        - "microarray_norm"
        - "metabolomics"
      data:
        type: "string"
        example: "\tSample 1\tSample2\tSample 3\nCD19\t10\t20\t5\nCD20\t4\t7\t30\n\
          MITF\t3\t9\t20\nPTEN\t10\t4\t20\n"
        description: "Tab-delimited expression matrix with the first column containing\
          \ gene / protein identifiers, the first row containing the sample labels\
          \ and each subsequent row corresponding to the expression of one gene in\
          \ all samples. The 'tab' character must be escaped using '\\t' and new-lines\
          \ must be escaped using '\\n'. If multiple datasets are submitted, shared\
          \ samples between the datasets must contain identical labels."
      design:
        $ref: "#/definitions/Design"
      parameters:
        type: "array"
        items:
          $ref: "#/definitions/Parameter"
    example:
      data: "\tSample 1\tSample2\tSample 3\nCD19\t10\t20\t5\nCD20\t4\t7\t30\nMITF\t\
        3\t9\t20\nPTEN\t10\t4\t20\n"
      design:
        comparison:
          group2: "Treatment"
          group1: "Control"
        analysisGroup:
        - "Treatment"
        - "Control"
        - "Treatment"
        samples:
        - "Sample 1"
        - "Sample 2"
        - "Sample 3"
      name: "First experiment"
      type: "rnaseq_counts"
      parameters:
      - name: "permutations"
        value: "10"
      - name: "permutations"
        value: "10"
  Parameter:
    type: "object"
    required:
    - "name"
    - "value"
    properties:
      name:
        type: "string"
        example: "permutations"
      value:
        type: "string"
        example: "10"
        description: "All parameter values must be converted to strings."
    example:
      name: "permutations"
      value: "10"
  Design:
    type: "object"
    required:
    - "analysisGroup"
    - "comparison"
    - "samples"
    properties:
      samples:
        type: "array"
        example:
        - "Sample 1"
        - "Sample 2"
        - "Sample 3"
        description: "List of all samples in the dataset. This will overwrite the\
          \ sample names (= column names) in the expression matrix. Therefore, the\
          \ number of entries in the array must match the number of samples in the\
          \ expression matrix."
        items:
          type: "string"
      comparison:
        $ref: "#/definitions/Design_comparison"
      analysisGroup:
        type: "array"
        example:
        - "Treatment"
        - "Control"
        - "Treatment"
        description: "Names the group to use for the main analysis"
        items:
          type: "string"
    description: "Specifies the experimental design for a dataset."
    example:
      comparison:
        group2: "Treatment"
        group1: "Control"
      analysisGroup:
      - "Treatment"
      - "Control"
      - "Treatment"
      samples:
      - "Sample 1"
      - "Sample 2"
      - "Sample 3"
    additionalProperties:
      type: "array"
      example:
      - "Lab 1"
      - "Lab 1"
      - "Lab 2"
      description: "Additional confounding factors can be specified as arrays following\
        \ the format of 'analysisGroup'. Not all GSA methods will make use of these\
        \ factors."
      items:
        type: "string"
...
  Design_comparison:
    required:
    - "group1"
    - "group2"
    properties:
      group1:
        type: "string"
        example: "Control"
      group2:
        type: "string"
        example: "Treatment"
    description: "Specifies the main comparison to perform in the analysis. 'group1'\
      \ and 'group2' refer to the labels specified in the 'analysisGroup' array."
    example:
      group2: "Treatment"
      group1: "Control"
...